ci: 정적 분석 도구 도입 및 ruff 보안/품질 규칙 확장
ruff lint 규칙에 S(보안), B(버그), A(builtins), C4(comprehensions),
SIM(simplify), PTH(pathlib), RUF(ruff-specific) 추가.
pip-audit(의존성 CVE 검사)와 radon(순환 복잡도/유지보수성) dev dependency 추가.

코드 수정:
- SIM108: if-else를 ternary로 변환 (cli.py, progress.py)
- SIM117: 중첩 with를 단일 with로 병합 (test_youtube.py)
- SIM102: 중첩 if를 단일 조건으로 병합 (auth.py)
- RUF046: 불필요한 int() 제거 (summary_generator.py)
- RUF022: __all__ 정렬 (youtube/__init__.py)
- RUF059/B007: 미사용 변수에 _ prefix (test_effects.py, test_scanner.py 등)
- S608 해소를 위해 f-string SQL을 문자열 연결로 변경 (cli.py)
- 기존 noqa 주석 제거, 불가피한 규칙은 pyproject.toml per-file-ignores로 관리

CI에 security-audit(pip-audit), code-quality(radon) 잡 추가 (총 7개 병렬).
pre-commit hook 스크립트 추가 (ruff check/format, mypy).

+++ tubearchive/utils/summary_generator.py
@@ -22,7 +22,7 @@ def format_timestamp(seconds: float) -> str:
     Returns:
         H:MM:SS 또는 M:SS 형식 문자열
     """
-    total_seconds = int(round(seconds))
+    total_seconds = round(seconds)
     hours = total_seconds // 3600
     minutes = (total_seconds % 3600) // 60
     secs = total_seconds % 60
